% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeAegypti.r
\docType{data}
\name{aeAegypti}
\alias{aeAegypti}
\title{Males and Females Aedes Aegypti data from Maletzke (2019)}
\format{
The data set aeAegypti is a data frame of 1800 observations of 9 variables. Each event is described by the wing beat frequency (wbf), and the frequencies of the first six harmonics obtained when either female or male \emph{Aedes Aegypti} mosquito cross an optical sensor's line-of-sigh. Both male (class = 2) and female (class = 1) of class factor.
}
\usage{
data(aeAegypti)
}
\description{
Contains events generated by a laser sensor to capture the flight dynamism of insects. It is a binary dataset compose by events from \emph{Aedes Aegypti} Female and Male.
}
\details{
The \code{aeAegypti} dataset is a subset of widely data collection effort involving more than one million instances from 20 different insect species. The dataset was collected varying the temperature and humidity. An observation is associated with a temperature range that varies from 23ºC to 35ºC.
}
\references{
Maletzke, A. G. (2019). Binary quantification in non-stationary scenarios. Doctoral Thesis, Instituto de Ciências Matemáticas e de Computação, University of São Paulo, São Carlos. Retrieved 2020-07-21, from www.teses.usp.br. <doi.org/10.11606/T.55.2020.tde-19032020-091709>

Moreira dos Reis, D., Maletzke, A., Silva, D. F., & Batista, G. E. (2018). Classifying and counting with recurrent contexts. In Proceedings of the 24th ACM SIGKDD International Conference on Knowledge Discovery & Data Mining (pp. 1983-1992). <doi.org/10.1145/3219819.3220059>
}
\author{
Andre Maletzke <andregustavom@gmail.com>
}
\keyword{datasets}
