% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{getHist}
\alias{getHist}
\title{Get a normalized histogram from score values.}
\usage{
getHist(scores, nbins)
}
\arguments{
\item{scores}{a numeric vector containing the scores.}

\item{nbins}{number of histogram bins.}
}
\value{
normalized histogram.
}
\description{
Get a normalized histogram from score values.
}
\examples{
positive_scores <- runif(200)^0.3
negative_scores <- 1 - runif(200)^0.3
getHist(positive_scores, 20)
getHist(negative_scores, 20)
}
\author{
Andre Maletzke <andregustavom@gmail.com>
}
