% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{TernarySearch}
\alias{TernarySearch}
\title{Implements the Ternary Search algorithm}
\usage{
TernarySearch(left, right, f, eps = 1e-04)
}
\arguments{
\item{left}{a numeric value containing the left bound.}

\item{right}{a numeric value containing the right bound.}

\item{f}{unimodal function.}

\item{eps}{stop criterium.}
}
\value{
the minimum of a unimodal function \code{f()} within \code{[left, right]}.
}
\description{
Ternary search is used for finding the best alpha value in DyS framework.
}
\author{
Andre Maletzke <andregustavom@gmail.com>
}
